Move ValidateToken to providers/internal_utils
